requirements.txt.txt
pandas
numpy
sqlalchemy
psycopg2-binary
streamlit
python-dotenv
os
datetime
plotly
sys
